Migrate to lndmobile
Lndmobile uses in-mem gRPC. Startup times are dramatically faster (at
least 50%), probably because there's no need to initialize a full
blown HTTPS server anymore.

Onward from this commit, we also make use of protobufjs to generate
typescript declaration files, so that we the correctly payload to the
gRPC API.

The downside of switching to in-mem gRPC is a more complex process of
getting data from lnd to the backend and back to lnd.
If the background service running lnd is another process, the main
activity cannot directly call the gRPC API. We make use of IPC with the
help of Android's Messenger API to deal with this issue.

Right now, there's also some issues figuring out what state lnd is in
(running or not?, wallet unlocked? etc). This has been loosely fixed
in our backend.

We also make use of a bound service now, so Android should never be able
to kill lnd without our app being killed as well.

Code for old backend is still available, but separated to different files.