Scheduled chain sync background job
This commit introduces a background job that periodically syncs the
blockchain. It binds to the LndMobileService which means that it will be
fine to open the app while the background job is running.

Right now the background job will be killed by Android if you would open
the app while it is running and then remove the app (from Recent Apps).
Android seems to rescheduled the job directly afterwards however.

It is using JetPack's WorkManager API to manage the job (using Job
Scheduler for new devices while maintaining backwards compatibility
using AlarmManager for old devices).

Right now the background job runs every 6 hours. The battery impact
seems to be minimal.
Even running at the minimally allowed interval (15 mins) seems to affect
the battery and load very little.

This commit also introduces a logging lib, HyperLog, to be able to save
log to accessible external storage.