Fix black screen on startup problem (attempt)
On Samsung devices, the app sometimes fails to start up resulting in a
black screen.

While the actual cause is still unclear, it has been confirmed that it's
related to the Scheduled Sync job and that it's sometimes started up
when the user starts the app (unknown why it starts up).
Why this causes a problem is still unknown as checkStatus in
LndMobile should handle any issues if both tries to start up Lnd.

This commit attempts to mitigate the problem by adding a `started` boolean
in `MainActivity` that's set when the app starts. The scheduled sync job
will then check if started is set to true and if it is stop all
operation and shut down.