Transaction log icons for known Lightning services 🔥
This commit introduces transaction log icons for a set of known
Lightning services. It works by checking the pubkey, webiste and/or
description of the invoice.

The images in question are are embedded inside the app.
They are (at the time of writing) 225x225 webp images with q set to 30.