android: Don't start TorService as a foreground service
This is due to a new rule in Android 14 where you have to declare
service type for each foreground service.
We don't have control over TorService and instead of creating
another foreground service that attaches to TorService, let's just
open it as a normal background binded service.

TorService should impliclitly live on because LndMobileService
is a foreground service bound to MainActivity and BlixtTor
is bound to MainActivity too.